Angular performance improvements (#2)
* Calculate Angular layout renderer child props via a pure pipe

Increase rendering and on change performance for all forms including nultiple layouts:
* Avoid unnecessary recalculation of the child render props.
* In turn, avoid some unnecessary change detection cycles in children.

* angular: Add service method to get global config

When querying the state from the `JsonFormsAngularService` via its `getState` method, the whole state is cloned.
This lead to a heavy performance penalty because the `JsonFormsAbstractControl` queried the state on every change detection cycle to get the global config. Thus, the whole state was cloned for every control on every change.
As the config is only a small part of the state, this problem can be alleviated by adding a getConfig method to the service that only clones the config.

* Extend `JsonFormsAngularService` with a `getConfig` method that returns a clone of the global config
* Use the new method in `JsonFormsAbstractControl`

Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>

* angular: Add service method to get locale

When querying the state from the `JsonFormsAngularService` via its `getState` method, the whole state is cloned.
This could lead to performance issues with the angular material renderer set because the `NumberControlRenderer` queried the state on every change detection cycle to get the locale. Thus, the whole state was cloned for every control of these types on every change.
To fix this, the `JsonFormsAngularService` now allows to query the locale directly without cloning the state

* Add method `getLocale` to `JsonFormsAngularService` that does not clone the state
* Use new method in `NumberControlRenderer`

Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>

Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>
Co-authored-by: Lucas Koehler <lkoehler@eclipsesource.com>